+++ src/priqueue.cpp
@@ -169,10 +169,10 @@ void PRIQUEUE::push(T* event)
 
 inline int log2_5(int n)
 {
-    const unsigned int b[] = { 0x2, 0xC, 0xF0, 0xFF00, 0xFFFF0000 };
-    const unsigned int S[] = { 1, 2, 4, 8, 16 };
+    const size_t b[] = { 0x2, 0xC, 0xF0, 0xFF00, 0xFFFF0000 };
+    const size_t S[] = { 1, 2, 4, 8, 16 };
 
-    unsigned int r = 0; // result of log2(v) will go here
+    size_t r = 0; // result of log2(v) will go here
     for (int i = 4; i >= 0; i--) // unroll for speed...   ...actually the compiler will unroll
     {
         if (n & b[i])
